# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AutoServis API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [work-orders]
      summary: List all work orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, waiting_parts, needs_return, completed]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [work-orders]
      summary: Create a new work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /work-orders/{id}:
    get:
      operationId: getWorkOrder
      tags: [work-orders]
      summary: Get work order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    patch:
      operationId: updateWorkOrder
      tags: [work-orders]
      summary: Update work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    delete:
      operationId: deleteWorkOrder
      tags: [work-orders]
      summary: Delete work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /work-orders/{id}/photos:
    get:
      operationId: listWorkOrderPhotos
      tags: [photos]
      summary: List photos for a work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
  /photos/{id}:
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    WorkOrder:
      type: object
      required: [id, licensePlate, status, createdAt]
      properties:
        id:
          type: integer
        vehicleId:
          type: ["integer", "null"]
        licensePlate:
          type: string
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, waiting_parts, needs_return, completed]
        km:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        oilChange:
          type: boolean
        transmissionOil:
          type: boolean
        brakes:
          type: boolean
        timing:
          type: boolean
        airFilter:
          type: boolean
        cabinFilter:
          type: boolean
        stk:
          type: boolean
        tireChange:
          type: boolean
        diagnostics:
          type: boolean
        lightsCheck:
          type: boolean
        brakeFluid:
          type: boolean
        frontAxleCheck:
          type: boolean
        rearAxleCheck:
          type: boolean
        frontShocksCheck:
          type: boolean
        rearShocksCheck:
          type: boolean
        geometry:
          type: boolean
        headlightAlignment:
          type: boolean
        serviceDate:
          type: ["string", "null"]
          description: Date when service was performed (for backdated/historical orders)
        otherWork:
          type: ["string", "null"]
        otherServices:
          type: ["string", "null"]
        laborHours:
          type: ["string", "null"]
          description: Number of labor hours (decimal as string, e.g. "2.5")
        laborPrice:
          type: ["integer", "null"]
          description: Labor price in CZK
        notes:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
    WorkOrderInput:
      type: object
      required: [licensePlate]
      properties:
        licensePlate:
          type: string
          minLength: 1
        km:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        oilChange:
          type: boolean
        transmissionOil:
          type: boolean
        brakes:
          type: boolean
        timing:
          type: boolean
        airFilter:
          type: boolean
        cabinFilter:
          type: boolean
        stk:
          type: boolean
        tireChange:
          type: boolean
        diagnostics:
          type: boolean
        lightsCheck:
          type: boolean
        brakeFluid:
          type: boolean
        frontAxleCheck:
          type: boolean
        rearAxleCheck:
          type: boolean
        frontShocksCheck:
          type: boolean
        rearShocksCheck:
          type: boolean
        geometry:
          type: boolean
        headlightAlignment:
          type: boolean
        serviceDate:
          type: ["string", "null"]
        otherWork:
          type: ["string", "null"]
        otherServices:
          type: ["string", "null"]
        laborHours:
          type: ["string", "null"]
        laborPrice:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    WorkOrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, waiting_parts, needs_return, completed]
        km:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        oilChange:
          type: boolean
        transmissionOil:
          type: boolean
        brakes:
          type: boolean
        timing:
          type: boolean
        airFilter:
          type: boolean
        cabinFilter:
          type: boolean
        stk:
          type: boolean
        tireChange:
          type: boolean
        diagnostics:
          type: boolean
        lightsCheck:
          type: boolean
        brakeFluid:
          type: boolean
        frontAxleCheck:
          type: boolean
        rearAxleCheck:
          type: boolean
        frontShocksCheck:
          type: boolean
        rearShocksCheck:
          type: boolean
        geometry:
          type: boolean
        headlightAlignment:
          type: boolean
        serviceDate:
          type: ["string", "null"]
        otherWork:
          type: ["string", "null"]
        otherServices:
          type: ["string", "null"]
        laborHours:
          type: ["string", "null"]
        laborPrice:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...
    Photo:
      type: object
      required: [id, workOrderId, url, createdAt]
      properties:
        id:
          type: integer
        workOrderId:
          type: integer
        url:
          type: string
        filename:
          type: string
        createdAt:
          type: string
          format: date-time
...